Clusters' cyclical start, end, mean and range

Function cluster.cyclo.pos computes indices of cluster position (start, end, mean and range) of cyclical-time-constrained clusters along different cycles.

cluster.cyclo.pos(x, k)

Arguments

x

A cyclical-time-constrained clustering object, as provided by function cyclo.hclust.

k

an integer scalar with the desired number of clusters (as in cutree).

Value

An object of class list, containing:

means: A table containing the mean position of each cluster in each cycle.

starts: A table containing the starting position of each cluster along each cycle.

ends: A table containing the ending position of each cluster along each cycle.

ranges: A table containing the ranges covered by each cluster within each cycle.

Details

A valuable aspect of a cyclical-time constrained clustering is the ability to compare clusters' position in several cycles. Displacement of clusters can be interpreted as advancing or delaying of a particular state along the cycles. The function cluster.cyclo.pos compute four indices allowing related to cluster position within cycles (see Value).

See also

Author

Nicolas Djeghri, UBO

Examples

#WRITE SOME R CODE IN THERE!!!